login signup function

+++ app/src/main/java/com/example/myapplication/dbhandler/MyDBSiteHandler.java
@@ -6,7 +6,6 @@ import android.content.Context;
 import android.database.Cursor;
 import android.database.sqlite.SQLiteDatabase;
 import android.database.sqlite.SQLiteOpenHelper;
-import android.util.Log;
 
 import androidx.annotation.Nullable;
 
@@ -22,6 +21,7 @@ public class MyDBSiteHandler extends SQLiteOpenHelper {
     public static final String COLUMN_ID = "_id";
     public static final String COLUMN_NAME = "url";
     public static final String COLUMN_TITLE = "title";
+    public static final String COLUMN_USER_ID = "userid";
 
     public MyDBSiteHandler(@Nullable Context context, @Nullable String name, @Nullable SQLiteDatabase.CursorFactory factory, int version) {
         super(context, DATABASE_NAME, factory, DATABASE_VERSION);
@@ -32,7 +32,8 @@ public class MyDBSiteHandler extends SQLiteOpenHelper {
         String query = "CREATE TABLE " + TABLE_SITES + "(" +
                 COLUMN_ID + " INTEGER PRIMARY KEY AUTOINCREMENT, " +
                 COLUMN_NAME + " TEXT, " +
-                COLUMN_TITLE + " TEXT " + ")";
+                COLUMN_TITLE + " TEXT, " +
+                COLUMN_USER_ID + " LONG " + ")";
         db.execSQL(query);
     }
 
@@ -42,24 +43,25 @@ public class MyDBSiteHandler extends SQLiteOpenHelper {
         onCreate(db);
     }
 
-    public void addUrl(Website website) {
+    public void addUrl(Website website, long id) {
         ContentValues values = new ContentValues();
         values.put(COLUMN_NAME, website.get_url());
         values.put(COLUMN_TITLE, website.getTitle());
+        values.put(COLUMN_USER_ID, id);
         SQLiteDatabase db = getWritableDatabase();
         db.insert(TABLE_SITES, null, values);
         db.close();
     }
 
-    public void deleteUrl(String urlName) {
+    public void deleteUrl(String urlName, long id) {
         SQLiteDatabase db = getWritableDatabase();
-        db.execSQL("DELETE FROM " + TABLE_SITES + " WHERE " + COLUMN_NAME + " =\"" + urlName + "\";");
+        db.execSQL("DELETE FROM " + TABLE_SITES + " WHERE " + COLUMN_NAME + " =\"" + urlName + "\" AND " + COLUMN_USER_ID + " = " + id + " ;");
     }
 
     @SuppressLint("Range")
-    public List<Website> databaseToString() {
+    public List<Website> databaseToString(long id) {
         SQLiteDatabase db = getWritableDatabase();
-        String query = "SELECT * FROM " + TABLE_SITES;
+        String query = "SELECT * FROM " + TABLE_SITES + " WHERE " + COLUMN_USER_ID + " = " + id;
 
         List<Website> websites = new ArrayList<>();
         Cursor c = db.rawQuery(query, null);
@@ -76,8 +78,8 @@ public class MyDBSiteHandler extends SQLiteOpenHelper {
         return websites;
     }
 
-    public void clearHistory() {
+    public void clearHistory(long id) {
         SQLiteDatabase db = getWritableDatabase();
-        db.execSQL("DELETE FROM " + TABLE_SITES);
+        db.execSQL("DELETE FROM " + TABLE_SITES + " WHERE " + COLUMN_USER_ID + " = " + id);
     }
 }
